+++ HISTORY.md
@@ -4,7 +4,8 @@
 
 ### 2020-07-22
 * morphic, blocks, gui: tweaked block-fading mouse-over
-* blocks, threads: tweaked context visualizations to be alpha-independent 
+* blocks, threads: tweaked context visualizations to be alpha-independent
+* gui: save block-transparency in  settings 
 
 ### 2020-07-21
 * blocks: tweaked block highlights for fade-out